@@ -1,5 +1,6 @@
 {
   "name": "@naturalcycles/sqlite-lib",
+  "type": "module",
   "scripts": {
     "prepare": "husky",
     "build": "dev-lib build",
@@ -9,19 +10,21 @@
     "lbt": "dev-lib lbt"
   },
   "dependencies": {
-    "@naturalcycles/db-lib": "^9.1.0",
-    "@naturalcycles/js-lib": "^14.27.0",
-    "@naturalcycles/nodejs-lib": "^13.1.3",
-    "@types/better-sqlite3": "^7.6.0",
-    "better-sqlite3": "^11.0.0",
-    "sql-template-strings": "^2.2.2",
-    "sqlite": "^5.0.1",
-    "sqlite3": "^5.1.1"
+    "@naturalcycles/db-lib": "^10",
+    "@naturalcycles/js-lib": "^14",
+    "@naturalcycles/nodejs-lib": "^13",
+    "@types/better-sqlite3": "^7",
+    "better-sqlite3": "^11",
+    "sql-template-strings": "^2",
+    "sqlite": "^5",
+    "sqlite3": "^5"
   },
   "devDependencies": {
-    "@naturalcycles/dev-lib": "^15.21.0",
-    "@types/node": "^22.7.5",
-    "jest": "^29.0.3"
+    "@naturalcycles/dev-lib": "^17",
+    "@types/node": "^22",
+    "@vitest/coverage-v8": "^3",
+    "tsx": "^4",
+    "vitest": "^3"
   },
   "files": [
     "dist",
@@ -41,7 +44,7 @@
     "url": "https://github.com/NaturalCycles/sqlite-lib"
   },
   "engines": {
-    "node": ">=20.13.0"
+    "node": ">=22.12.0"
   },
   "version": "1.2.1",
   "description": "CommonDB implementation based on SQLite",
