@@ -1,4 +1,10 @@
-// prettier-ignore
-module.exports = [
-  ...require('@naturalcycles/dev-lib/cfg/eslint.config'),
+import sharedConfig from '@naturalcycles/dev-lib/cfg/eslint.config.js'
+
+export default [
+  ...sharedConfig,
+  {
+    rules: {
+      '@typescript-eslint/consistent-type-imports': 2,
+    },
+  },
 ]
