@@ -1,12 +1,14 @@
-import {
+import type {
   CommonDBCreateOptions,
   CommonKeyValueDB,
-  commonKeyValueDBFullSupport,
   IncrementTuple,
   KeyValueDBTuple,
 } from '@naturalcycles/db-lib'
-import { AppError, CommonLogger } from '@naturalcycles/js-lib'
-import { boldWhite, readableCreate, ReadableTyped } from '@naturalcycles/nodejs-lib'
+import { commonKeyValueDBFullSupport } from '@naturalcycles/db-lib'
+import type { CommonLogger } from '@naturalcycles/js-lib'
+import { AppError } from '@naturalcycles/js-lib'
+import type { ReadableTyped } from '@naturalcycles/nodejs-lib'
+import { boldWhite, readableCreate } from '@naturalcycles/nodejs-lib'
 import type { Database, Options } from 'better-sqlite3'
 import BetterSqlite3 from 'better-sqlite3'
 
