@@ -1,6 +1,6 @@
 import { Readable } from 'node:stream'
-import { ReadableTyped } from '@naturalcycles/nodejs-lib'
-import { Database, Statement } from 'sqlite'
+import type { ReadableTyped } from '@naturalcycles/nodejs-lib'
+import type { Database, Statement } from 'sqlite'
 
 /**
  * Based on: https://gist.github.com/rmela/a3bed669ad6194fb2d9670789541b0c7
