@@ -1,4 +1,8 @@
-import * as path from 'node:path'
+import { dirname, join } from 'node:path'
+import { fileURLToPath } from 'node:url'
 
-export const projectDir = path.join(`${__dirname}/../..`)
+const __filename = fileURLToPath(import.meta.url)
+const __dirname = dirname(__filename)
+
+export const projectDir = join(`${__dirname}/../..`)
 export const tmpDir = `${projectDir}/tmp`
